### 2018-02-18 #587 @whedon
[PRE REVIEW]: Flux: Seamless machine learning with Julia
**Submitting author:** @MikeInnes (<a href="http://orcid.org/0000-0003-0788-0242">Michael Innes</a>)
**Repository:** <a href="https://github.com/FluxML/Flux.jl" target ="_blank">https://github.com/FluxML/Flux.jl</a>
**Version:** v0.4.1
