### 2018-02-20 e6f5564
Convert OneHot CuArrays to dense CuArrays before passing to CUDNN methods
