+++ src/treelike.jl
@@ -10,7 +10,7 @@ function treelike(T, fs = fieldnames(T))
   @eval current_module() begin
     children(x::$T) = ($([:(x.$f) for f in fs]...),)
     mapchildren(f, x::$T) = $T(f.(children(x))...)
-    adapt(T, x::$T) = mapchildren(x -> adapt(T, x), x)
+    adapt(T, x::$T) = mapleaves(x -> adapt(T, x), x)
   end
 end
 
