+++ src/treelike.jl
@@ -21,9 +21,6 @@ function mapleaves(f, x; cache = ObjectIdDict())
   cache[x] = isleaf(x) ? f(x) : mapchildren(x -> mapleaves(f, x, cache = cache), x)
 end
 
-export mapparams
-@deprecate mapparams(f, x) mapleaves(f, x)
-
 using DataFlow: OSet
 
 function prefor(f, x; seen = OSet())
