### 2018-02-25 719bea3
slight generalization of cat.
