+++ test/tracker.jl
@@ -31,6 +31,13 @@ gradtest(f, dims...) = gradtest(f, rand.(dims)...)
 @test gradtest(vcat, rand(5), rand(3), rand(8))
 @test gradtest(vcat, rand(5,2), rand(3,2), rand(8,2))
 
+@test gradtest((i...) -> cat(1,i...), rand(5), rand(3))
+@test gradtest((i...) -> cat(1,i...), rand(5), rand(8))
+@test gradtest((i...) -> cat(1,i...), rand(5,2),rand(3,2), rand(8,2))
+@test gradtest((i...) -> cat(2,i...), rand(5,1), rand(5,1))
+@test gradtest((i...) -> cat(2,i...), rand(5,1), rand(5,4))
+@test gradtest((i...) -> cat(2,i...), rand(5,2),rand(5,4), rand(5,8))
+
 @test gradtest(kron,rand(5), rand(3))
 @test gradtest(kron, rand(5), rand(3), rand(8))
 @test gradtest(kron,rand(5,1), rand(3,1))
