### 2018-02-26 #587 @MikeInnes
I could also suggest @trallard, who I know has been involved with Julia. (Hope it's OK to ping)
