+++ test/tracker.jl
@@ -1,6 +1,6 @@
 using Flux.Tracker, Base.Test, NNlib
 using Flux.Tracker: TrackedReal, gradcheck
-using NNlib
+using NNlib: conv
 
 gradtest(f, xs::AbstractArray...) = gradcheck((xs...) -> sum(sin.(f(xs...))), xs...)
 gradtest(f, dims...) = gradtest(f, rand.(dims)...)
@@ -60,9 +60,9 @@ end
   2y + x
 end
 
-@test gradtest(conv2d, rand(10, 10, 3, 2), randn(2, 2, 3, 2))
-@test gradtest(x -> maxpool2d(x, 2), rand(10, 10, 3, 2))
-@test gradtest(x -> avgpool2d(x, 2), rand(10, 10, 3, 2))
+@test gradtest(conv, rand(10, 10, 3, 2), randn(2, 2, 3, 2))
+@test gradtest(x -> maxpool(x, (2,2)), rand(10, 10, 3, 2))
+@test gradtest(x -> meanpool(x, (2,2)), rand(10, 10, 3, 2))
 
 @test (param([1,2,3]) .< 2) == [true, false, false]
 
