### 2018-02-26 466b5c5
cpu/gpu conveniences
