+++ src/Flux.jl
@@ -10,7 +10,7 @@ using MacroTools: @forward
 export Chain, Dense, RNN, LSTM, GRU, Conv, Conv2D,
   Dropout, LayerNorm, BatchNorm,
   SGD, ADAM, Momentum, Nesterov, AMSGrad,
-  param, params, mapleaves
+  param, params, mapleaves, cpu, gpu
 
 @reexport using NNlib
 
