### 2018-02-27 bdb8aae
move cache logic
