+++ src/jit/JIT.jl
@@ -1,7 +1,6 @@
 module JIT
 
 include("shapes.jl")
-include("inplace.jl")
 include("trace.jl")
 include("lib.jl")
 
