+++ src/jit/shapes.jl
@@ -35,3 +35,17 @@ function restructure(sh::Shape{T}, buf::Vector{UInt8}) where T
   buf = unsafe_wrap(Array, pointer(buf), sizeof(sh))
   reshape(reinterpret(T, buf), size(sh))
 end
+
+# Execution with caches
+
+mutable struct Cached{F,A}
+  f::F
+  buffer::A
+end
+
+function (c::Cached)(args...)
+  sh = shape(c.f, shape(args)...)
+  bytes(sh) > length(c.buffer) && (c.buffer = similar(c.buffer, bytes(sh)))
+  y = restructure(sh, c.buffer)
+  inplace!(c.f, y, args...)
+end
