+++ src/jit/trace.jl
@@ -2,7 +2,8 @@
 
 using ..Flux.Tracker: Tracked, Broadcasted, param, tracker, istracked, isleaf
 using DataFlow
-using DataFlow: inputnode, constant
+using DataFlow: Call, Lambda, iscall, isconstant, prewalk, vertex, syntax,
+  inputnode, constant
 
 vcall(f, args...) = vertex(DataFlow.Call(), constant(f), args...)
 vcall(f::Broadcasted, args...) = vcall(broadcast, constant(f.f), args...)
@@ -22,4 +23,24 @@ end
 function trace(f, args...)
   inputs = param.(args)
   graph(f(inputs...), inputs...)
+end
+
+# Graph manipulation
+
+function cacheall(v, buf = () -> UInt8[])
+  prewalk(v) do v
+    iscall(v) && isconstant(v[1]) || return v
+    f = v[1].value.value
+    return vertex(Call(), constant(Cached(f, buf())), v[2:end]...)
+  end
+end
+
+function eval_func(v, n)
+  v = vertex(Lambda(n, v))
+  v |> syntax |> eval
+end
+
+function compile(f, args...)
+  v = trace(f, args...)
+  eval_func(cacheall(v), length(args))
 end
