### 2018-02-28 3ac6a8e
n-d conv tests
