+++ test/tracker.jl
@@ -60,7 +60,9 @@ end
   2y + x
 end
 
+@test gradtest(conv, rand(10, 3, 2), randn(2, 3, 2))
 @test gradtest(conv, rand(10, 10, 3, 2), randn(2, 2, 3, 2))
+@test gradtest(conv, rand(10, 10, 10, 3, 2), randn(2, 2, 2, 3, 2))
 @test gradtest(x -> maxpool(x, (2,2)), rand(10, 10, 3, 2))
 @test gradtest(x -> meanpool(x, (2,2)), rand(10, 10, 3, 2))
 
