### 2018-02-28 #587 @MikeInnes
Alternatively @ysimillides may be able to review.
