### 2018-02-28 4606339
nd pooling tests
