### 2018-02-28 7606b1a
single-batch convolution
