### 2018-02-28 ccef9f4
jit softmax
