+++ src/jit/JIT.jl
@@ -1,5 +1,7 @@
 module JIT
 
+using MacroTools
+
 include("shapes.jl")
 include("trace.jl")
 include("lib.jl")
