### 2018-02-28 2eb38ee
update gpu api
