+++ test/cuda/cuda.jl
@@ -21,6 +21,11 @@ cm = gpu(m)
 @test all(p isa TrackedArray && p.data isa CuArray for p in params(cm))
 @test cm(gpu(rand(10, 10))) isa TrackedArray{Float32,2,CuArray{Float32,2}}
 
+# Fails in Pkg.test ffs
+# c = gpu(Conv((2,2),3=>4))
+# l = c(gpu(rand(10,10,3,2)))
+# Flux.back!(sum(l))
+
 end
 
 CuArrays.cudnn_available() && include("cudnn.jl")
