+++ src/Flux.jl
@@ -13,6 +13,7 @@ export Chain, Dense, RNN, LSTM, GRU, Conv, Conv2D,
   param, params, mapleaves, cpu, gpu
 
 @reexport using NNlib
+using NNlib: @fix
 
 include("tracker/Tracker.jl")
 using .Tracker
