### 2018-03-01 8019f78
use normal log
