### 2018-03-02 19f691d
Use `cache.julialang.org` to store ML models
It's annoying that when third party servers go down our tests break.
Let's at least make sure that if our tests break due to server outages
it's our fault.
