### 2018-03-03 7c6c797
Combine tests
