### 2018-03-04 9acfdbd
Dense maxout
