### 2018-03-05 #602 @Kevin-Mattheus-Moerman
@ysimillides @vchuravy thanks for reviewing this submission :rocket:  I've created tick boxes for the both of you at the top of this issue. You can leave review comments here but for larger items we suggest you open an issue on the project repo and link to it here.
Here are our reviewer guidelines: http://joss.theoj.org/about#reviewer_guidelines

Let me know if you have questions.

Let the reviewing begin :tada:
