### 2018-03-05 #602 @Kevin-Mattheus-Moerman
@davidbarber @taheris are you interested in reviewing [this submission](https://github.com/FluxML/Flux.jl) for [JOSS](http://joss.theoj.org/) as well? Review is focused largely on the software and also a [short paper :bookmark: ](https://github.com/openjournals/joss-papers/blob/joss.00602/joss.00602/10.21105.joss.00602.pdf)
