### 2018-03-05 fa2717f
Add `hcat()` for tracked arrays
