### 2018-03-05 #602 @Kevin-Mattheus-Moerman
Great the more the better. @ChrisRackauckas I'll create a check list for you at the top of this issue as well.
