### 2018-03-05 5c7f856
cpu docs
