### 2018-03-05 FluxML/Flux.jl#195 @ysimillides
Community guidelines inclusion
As per the community guidelines section in https://github.com/openjournals/joss-reviews/issues/602  (  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support) it may be helpful if these where explicitly mentioned in the readme / contributing.md file (or a link to the community page found here https://github.com/FluxML/Flux.jl/blob/master/docs/src/community.md).
