### 2018-03-06 c95a97f
make epochs available
