### 2018-03-06 eab26be
optimiser state
