+++ versions.js
@@ -5,6 +5,7 @@ var DOC_VERSIONS = [
   "release-0.3",
   "release-0.2",
   "release-0.1",
+  "v0.5.0",
   "v0.4.1",
   "v0.4.0",
   "v0.3.4",
