### 2018-03-06 86d782a
implement `prod` for `TrackedArray`
