### 2018-03-06 3babeeb
scalar hashing
