### 2018-03-06 6445295
Better download detection
