+++ src/treelike.jl
@@ -8,8 +8,8 @@ mapchildren(f, x::Tuple) = map(f, x)
 
 function treelike(T, fs = fieldnames(T))
   @eval current_module() begin
-    children(x::$T) = ($([:(x.$f) for f in fs]...),)
-    mapchildren(f, x::$T) = $T(f.(children(x))...)
+    Flux.children(x::$T) = ($([:(x.$f) for f in fs]...),)
+    Flux.mapchildren(f, x::$T) = $T(f.($children(x))...)
   end
 end
 
