### 2018-03-06 36baa7e
convnet primitives
