### 2018-03-07 7c72147
add gradient check for `prod` and fix `dims` in `back(::typeof(prod),...)`
