### 2018-03-07 c00f7f8
implement `back` of `repmat`
