+++ test/tracker.jl
@@ -32,6 +32,9 @@ gradtest(f, dims...) = gradtest(f, rand.(dims)...)
 @test gradtest(vcat, rand(5,2), rand(3,2), rand(8,2))
 @test gradtest(x -> permutedims(x, [3,1,2]), rand(4,5,6))
 
+@test gradtest(x -> repmat(x, 5,5), rand(4,5))
+@test gradtest(x -> repmat(x, 5), rand(4,5))
+
 @test gradtest(kron,rand(5), rand(3))
 @test gradtest(kron, rand(5), rand(3), rand(8))
 @test gradtest(kron,rand(5,1), rand(3,1))
