### 2018-03-07 43af389
change `prod` implementation to avoid small xs
