### 2018-03-07 9ccbac8
jit gpu support
