### 2018-03-08 eaae552
Use dynamic eps
