### 2018-03-08 #602 @ChrisRackauckas
I think I will need to step down after reading the conflict of interests policy. I am currently (well just started) working with @MikeInnes on a publication which utilizes Flux.jl and so someone with less experience of the software, it's community, and its authors can probably do a more impartial scientific assessment, though I will still go through and open docs PRs where I see fit.
