### 2018-03-08 #602 @Kevin-Mattheus-Moerman
Thanks @ChrisRackauckas that sounds good
