+++ src/tracker/scalar.jl
@@ -61,7 +61,7 @@ for (M, f, arity) in DiffRules.diffrules()
   end
 end
 
-# Eliminating ambiguous
+# Eliminating ambiguity
 import Base:^
 
 ^(a::TrackedReal, b::Integer) = track(^, a, b)
