### 2018-03-16 db2d9ef
Update BatchNorm documentation
