### 2018-03-16 07b0f95
Tests for batch norm with 2D and 3D convolutions.
