### 2018-03-16 e0cda6f
Added CIFAR10 dataset.
