### 2018-03-16 FluxML/Flux.jl#195 @MikeInnes
I've updated the readme, does that clear things up?
