### 2018-03-16 #602 @MikeInnes
I've tried to address @ysimillides comments, he can say if that's sufficient.

As no other reviewers have stepped up over a month, seems like it might be challenging to get more review on this. Is there any chance of moving forward with the one review?
