### 2018-03-16 FluxML/Flux.jl#195 @ysimillides
Yes, thats perfectly fine by me :) . I'll also confirm on the review page.
