### 2018-03-16 1c5f8e3
ndims for shapes
