### 2018-03-17 b415935
Update test, show and docstring
