### 2018-03-18 4703b7b
Added Local Contrast Normalization
