### 2018-03-19 07ed439
minor update for gpu.md code block
