+++ docs/src/gpu.md
@@ -55,7 +55,7 @@ Tracked 5-element CuArray{Float32,1}:
 
 The analogue `cpu` is also available for moving models and data back off of the GPU.
 
-```
+```julia
 julia> x = rand(10) |> gpu
 10-element CuArray{Float32,1}:
  0.235164
@@ -67,4 +67,4 @@ julia> x |> cpu
  0.235164
  ⋮
  0.192538
- ```
+```
