### 2018-03-19 c724237
added stride for pooling in tracker
