### 2018-03-19 5cc6813
added stride for pooling in tracker
