### 2018-03-20 5a3457c
Local_Contrast_Norm name changed
