### 2018-03-20 572df2b
Local_Contrast_Norm name changed
