### 2018-03-23 35299d4
Fix type instability when loading onto a gpu.
Also fixes Issue #216.
