### 2018-03-23 b9a66c6
Fix error in initialization of σ.
