### 2018-03-27 #602 @Kevin-Mattheus-Moerman
@MikeInnes I've reviewed the paper. About the statement

> Flux has been used heavily for natural language processing...

would you be able to add one or more references for this?
