+++ docs/src/training/training.md
@@ -29,31 +29,27 @@ loss(x, y) = Flux.mse(m(x), y)
 Flux.train!(loss, data, opt)
 ```
 
-The objective will almost always be defined in terms of some *cost function* that measures the distance of the prediction `m(x)` from the target `y`. Flux has several of these built in, like `mse` for mean squared error or `crossentropy` for cross entropy loss, but you can calculate it however you want.
+In supervised tasks, the objective is tipically defined in terms of some *cost function* that measures the distance of the prediction `m(x)` from the target `y`. Flux has several of these built in, like `mse` for mean squared error or `crossentropy` for cross entropy loss, but you can calculate it however you want. In general scenarios, the loss function can take any number of inputs.
 
-## Datasets
+## Data
 
-The `data` argument provides a collection of data to train with (usually a set of inputs `x` and target outputs `y`). For example, here's a dummy data set with only one data point:
-
-```julia
-x = rand(784)
-y = rand(10)
-data = [(x, y)]
-```
-
-`Flux.train!` will call `loss(x, y)`, calculate gradients, update the weights and then move on to the next data point if there is one. We can train the model on the same data three times:
+The data argument `train!` provides a collection of data to train with (usually a set of inputs `x` and target outputs `y`). 
+For large datasets it is often convenient to arrange the data in small `batches`.
+For example, here's a synthetic data set with 10 data points partitioned in minibatches of size 1:
 
 ```julia
-data = [(x, y), (x, y), (x, y)]
-# Or equivalently
-data = Iterators.repeated((x, y), 3)
+X = rand(784, 10)
+Y = rand(10)
+data = batches(X, Y, batchsize=1)
 ```
+
+For each `(x, y)` in `data`, `Flux.train!` will call `loss(x, y)`, calculate gradients, update the weights and then move on to the next data point if there is one.
 
 It's common to load the `x`s and `y`s separately. In this case you can use `zip`:
 
 ```julia
 xs = [rand(784), rand(784), rand(784)]
-ys = [rand( 10), rand( 10), rand( 10)]
+ys = [2, 3, 1]
 data = zip(xs, ys)
 ```
 
