### 2018-04-02 2d8d390
rework loss functions
