+++ src/Flux.jl
@@ -29,6 +29,8 @@ include("onehot.jl")
 include("treelike.jl")
 
 include("layers/stateless.jl")
+export cross_entropy, bce, mse, bce_with_logits, nll
+
 include("layers/basic.jl")
 include("layers/conv.jl")
 include("layers/recurrent.jl")
