### 2018-04-02 8b955c4
fixes and tests
