+++ docs/src/models/regularisation.md
@@ -8,14 +8,14 @@ For example, say we have a simple regression.
 
 ```julia
 m = Dense(10, 5)
-loss(x, y) = crossentropy(softmax(m(x)), y)
+loss(x, y) = cross_entropy(m(x), y)
 ```
 
 We can regularise this by taking the (L2) norm of the parameters, `m.W` and `m.b`.
 
 ```julia
 penalty() = vecnorm(m.W) + vecnorm(m.b)
-loss(x, y) = crossentropy(softmax(m(x)), y) + penalty()
+loss(x, y) = cross_entropy(m(x), y) + penalty()
 ```
 
 When working with layers, Flux provides the `params` function to grab all
@@ -37,9 +37,9 @@ Here's a larger example with a multi-layer perceptron.
 m = Chain(
   Dense(28^2, 128, relu),
   Dense(128, 32, relu),
-  Dense(32, 10), softmax)
+  Dense(32, 10))
 
-loss(x, y) = crossentropy(m(x), y) + sum(vecnorm, params(m))
+loss(x, y) = cross_entropy(m(x), y) + sum(vecnorm, params(m))
 
 loss(rand(28^2), rand(10))
 ```
