### 2018-04-02 d582e02
-extend crossentropy to matrices -reorg files
