+++ src/Flux.jl
@@ -8,7 +8,6 @@ using Juno, Requires, Reexport
 using MacroTools: @forward
 
 export Chain, Dense, RNN, LSTM, GRU, Conv, Conv2D,
-  Dropout, LayerNorm, BatchNorm,
   SGD, ADAM, Momentum, Nesterov, AMSGrad,
   param, params, mapleaves, cpu, gpu
 
@@ -28,13 +27,15 @@ include("utils.jl")
 include("onehot.jl")
 include("treelike.jl")
 
-include("layers/stateless.jl")
+include("layers/loss.jl")
 export cross_entropy, bce, mse, bce_logit, nll
 
 include("layers/basic.jl")
 include("layers/conv.jl")
 include("layers/recurrent.jl")
+
 include("layers/normalisation.jl")
+export Dropout, LayerNorm, BatchNorm, normalise
 
 include("data/Data.jl")
 
