+++ test/runtests.jl
@@ -7,7 +7,7 @@ srand(0)
 include("utils.jl")
 include("tracker.jl")
 include("layers/normalisation.jl")
-include("layers/stateless.jl")
+include("layers/loss.jl")
 include("optimise.jl")
 include("data.jl")
 include("jit.jl")
