+++ src/layers/normalisation.jl
@@ -140,3 +140,15 @@ function Base.show(io::IO, l::BatchNorm)
   (l.λ == identity) || print(io, ", λ = $(l.λ)")
   print(io, ")")
 end
+
+"""
+    normalise(x::AbstractVecOrMat)
+
+Normalise each column of `x` to mean 0 and standard deviation 1.
+"""
+function normalise(x::AbstractVecOrMat)
+  μ = mean(x, 1)
+  σ = std(x, 1, mean = μ)
+  return (x .- μ) ./ σ
+end
+
