### 2018-04-02 480bce1
merge onehotbatch into onehot
