### 2018-04-02 ea9b547
NADAM optimizer
