+++ src/Flux.jl
@@ -36,8 +36,6 @@ include("layers/normalisation.jl")
 
 include("data/Data.jl")
 
-include("jit/JIT.jl")
-
 @require CuArrays include("cuda/cuda.jl")
 
 end # module
