+++ src/jit/JIT.jl
@@ -1,9 +1,0 @@
-module JIT
-
-using MacroTools
-
-include("shapes.jl")
-include("trace.jl")
-include("lib.jl")
-
-end
