### 2018-04-14 #602 @MikeInnes
I've updated the paper a little to reference some of the NLP work that's out there via Minibatch.jl and the model zoo. It's a bit tricky that a lot of the really neat stuff is either industrial or in preparation / review right now (indeed a big part of the motivation for this submission is so that they can cite us :)
