+++ README.md
@@ -38,7 +38,7 @@ b = param(randn(2))
 y(x) = σ.(W * x .+ b)
 ```
 
-If that's *still* not enough, you can go as deep as you want, even writing your own CUDA kernels! All this can be freely mixed-and-matched in a single model or script, and it all runs interactively via Jupyter or Juno.
+If that's *still* not enough, you can go as deep as you want, even writing your own CUDA kernels with [CUDAnative](https://github.com/JuliaGPU/CUDAnative.jl)! All this can be freely mixed-and-matched in a single model or script, and it all runs interactively via Jupyter or Juno.
 
 ```julia
 function gpu_add(a, b, c)
