### 2018-04-15 f6097d5
Scalar pad/stride for Conv constructor
