+++ src/layers/conv.jl
@@ -24,11 +24,16 @@ Conv(w::AbstractArray{T}, b::AbstractVector{T}, σ = identity;
        stride = 1, pad = 0) where T =
   Conv(σ, w, b, stride, pad)
 
+#Conv(k::NTuple{N,Integer}, ch::Pair{<:Integer,<:Integer}, σ = identity; init = initn,
+#     stride::NTuple{N,Integer} = map(_->1,k),
+#     pad::NTuple{N,Integer} = map(_->0,k)) where N =
+#  Conv(param(init(k..., ch...)), param(zeros(ch[2])), σ,
+#       stride = stride, pad = pad)
+
 Conv(k::NTuple{N,Integer}, ch::Pair{<:Integer,<:Integer}, σ = identity; init = initn,
-     stride::NTuple{N,Integer} = map(_->1,k),
-     pad::NTuple{N,Integer} = map(_->0,k)) where N =
+     stride::Integer = 1, pad::Integer = 0) where N =
   Conv(param(init(k..., ch...)), param(zeros(ch[2])), σ,
-       stride = stride, pad = pad)
+       stride = map(_->stride,k), pad = map(_->pad,k))
 
 Flux.treelike(Conv)
 
