+++ src/Flux.jl
@@ -32,7 +32,7 @@ include("layers/stateless.jl")
 include("layers/basic.jl")
 include("layers/conv.jl")
 include("layers/recurrent.jl")
-include("layers/normalisation.jl")
+include("layers/normalise.jl")
 
 include("data/Data.jl")
 
