### 2018-04-15 b080f5c
Scalar pad and stride
