### 2018-04-15 2f5473d
added expand in conv constructor
