### 2018-04-15 d12fb98
nicer batchnorm shape error
