### 2018-04-17 8f73dc6
fix gpu cross entropy
