### 2018-04-17 baff205
gpu broadcast fix
