### 2018-04-18 5faf5da
doc: add GRU to layer reference
