### 2018-04-18 5aba4c8
layers: fix docstring of LSTM and GRU
[ci skip]
