### 2018-04-19 c5c23b6
Mention CUDAnative.jl's install instructions
