### 2018-04-19 4076670
replace CUDAnative.jl with GPU support doc mention
