### 2018-04-19 e5c9486
add CUDAnative.jl mention
