### 2018-04-20 eaaf5fd
vcat arrays with ndims>2
