### 2018-04-24 #602 @Kevin-Mattheus-Moerman
@MikeInnes thanks for updating the paper. BTW if works are in submission and you'd like to cite them you could cite the DOI of a pre-print for instance.
