### 2018-04-24 #602 @MikeInnes
I'll poke those folks in case there are any public DOIs around that I can cite, but if this paper can go through before then I'm equally happy to leave it.
