### 2018-04-26 b6508e2
add adamax
