### 2018-04-26 4586bda
export/test adamax
