### 2018-04-26 8c042bd
element wise max()
