### 2018-04-27 acf0a79
Add stopping conditions and iter_throttle
