+++ src/Flux.jl
@@ -11,6 +11,7 @@ export Chain, Dense, RNN, LSTM, GRU, Conv, Conv2D,
   Dropout, LayerNorm, BatchNorm,
   SGD, ADAM, Momentum, Nesterov, AMSGrad,
   param, params, mapleaves, cpu, gpu
+  
 
 @reexport using NNlib
 using NNlib: @fix
