### 2018-04-27 f1eec0e
CIFAR-10 Dataset
Added the CIFAR-10 Dataset under the Flux.Data Module. This allows users to directly use the CIFAR-10 Dataset similar to the MNIST Dataset.
