### 2018-04-27 cfd29b9
Backpropagation for `maximum` and `minimum`
