### 2018-04-28 bc1d422
Update CIFAR-10
Modify labelnames() function to directly use inbuilt functions
