### 2018-05-01 7d7d895
rm this deprecation for 0.6
