### 2018-05-02 1c189c6
cat with multiple dims #156
Co-authored-by: americast <sayan.sinha@iitkgp.ac.in>
