### 2018-05-02 5fc6190
RowVector tests
