### 2018-05-02 59324c0
hcat tests #194
Co-authored-by: Elliot Saba <staticfloat@gmail.com>
