### 2018-05-02 94bb064
more tests of array promotion for concatenation
# Conflicts:
#	test/tracker.jl
