### 2018-05-02 13daaec
Refactored tests
