### 2018-05-02 bcef5c4
Support hcat and cat
