### 2018-05-03 #602 @Kevin-Mattheus-Moerman
@MikeInnes thanks, all boxes are ticked and it looks like we are all set. I recommend acceptance of this submission.
At this point, can you please make an archive of the final reviewed software in [Zenodo](https://zenodo.org/), or [figshare](https://figshare.com/), or another service, and update this thread with the DOI of the archive? Once we have that available we can process acceptance.
