### 2018-05-03 #602 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
